requirements.txt
Django==1.11.5
requests
pyOpenSSL==17.3.0